backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.3
python-decouple>=3.8
stripe>=9.12
django-filter>=24.1
requests>=2.32
psycopg2-binary>=2.9
